完善选择封面，对CMTime处理

+++ Alpface/Camera/AlpVideoCamera/CustomVideoCamera/AlpVideoCameraCoverSlider.h
@@ -11,26 +11,26 @@
 NS_ASSUME_NONNULL_BEGIN
 
 typedef struct _AlpVideoCameraCoverSliderRange {
-    CGFloat location;
-    CGFloat length;
+    NSInteger location;
+    NSInteger length;
 } AlpVideoCameraCoverSliderRange;
 
-NS_INLINE AlpVideoCameraCoverSliderRange AlpVideoCameraCoverSliderMakeRange(CGFloat loc,
-                                                                            CGFloat len) {
+NS_INLINE AlpVideoCameraCoverSliderRange AlpVideoCameraCoverSliderMakeRange(NSInteger loc,
+                                                                            NSInteger len) {
     AlpVideoCameraCoverSliderRange r;
     r.location = loc;
     r.length = len;
     return r;
 }
 
-NS_INLINE CGFloat AlpVideoCameraCoverSliderMaxRange(AlpVideoCameraCoverSliderRange range) {
+NS_INLINE NSInteger AlpVideoCameraCoverSliderMaxRange(AlpVideoCameraCoverSliderRange range) {
     return (range.location + range.length);
 }
 
 IB_DESIGNABLE
 @interface AlpVideoCameraCoverSlider : UIControl
 
-@property (nonatomic) IBInspectable CGFloat maximumValue;
+@property (nonatomic) IBInspectable NSInteger maximumValue;
 @property (nonatomic) IBInspectable UIImage *thumbImage;
 @property (nonatomic) IBInspectable AlpVideoCameraCoverSliderRange range;
 @property (nonatomic, readonly) UIView *rangeThumbView;
